Rename NoneSignScanner
